add pdf folder
